requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
pyparsing==3.0.6
matplotlib>=3.2.2
numpy==1.19.5
opencv-python==4.5.4.60
Pillow==8.4.0
PyYAML==6.0
requests==2.26.0
scipy==1.7.3
tqdm==4.62.3
cmake==3.18.2

#torch>=1.7.1
#torchvision>=0.8.2
#dlib>=19.21.0

####################################################### CUDA ##############################################################
##   	conda install pytorch==1.7.1 torchvision==0.8.2 torchaudio==0.7.2 cudatoolkit=10.1 -c pytorch                    ##
## pip3 install torch==1.7.1+cu101 torchvision==0.8.2+cu101 -f https://download.pytorch.org/whl/cu101/torch_stable.html  ##
##      https://www.linkedin.com/pulse/installing-dlib-cuda-support-windows-10-chitransh-mundra	                         ##
###########################################################################################################################


# plotting ------------------------------------

pandas==1.3.5
seaborn==0.11.0

# deep_sort -----------------------------------

easydict